+++ projectroles/tests/test_views_api.py
@@ -11,7 +11,6 @@ from django.urls import reverse
 from django.utils import timezone
 
 from knox.models import AuthToken
-
 from test_plus.test import APITestCase
 
 from projectroles import views_api
@@ -64,7 +63,7 @@ INVITE_USER_EMAIL = 'new1@example.com'
 INVITE_USER2_EMAIL = 'new2@example.com'
 INVITE_MESSAGE = 'Message'
 
-#: Special value to use for empty knox token
+# Special value to use for empty knox token
 EMPTY_KNOX_TOKEN = '__EmpTy_KnoX_tOkEn_FoR_tEsT_oNlY_0xDEADBEEF__'
 
 
@@ -99,7 +98,8 @@ class SODARAPIViewTestMixin:
 
         :param user: User object
         :param full_result: Return full result of AuthToken creation if True
-        :return: Token string or AuthToken creation tuple (``EMPTY_KNOX_TOKEN`` if user is ``None``)
+        :return: Token string or AuthToken creation tuple (EMPTY_KNOX_TOKEN if
+                 user is None)
         """
         if user is None:
             return EMPTY_KNOX_TOKEN
@@ -160,7 +160,7 @@ class SODARAPIViewTestMixin:
         Return auth header based on token.
 
         :param token: Token string
-        :return: Dict, empty if ``token`` is ``None``
+        :return: Dict, empty if token is None
         """
         if token is EMPTY_KNOX_TOKEN:
             return {}